// src/api/generated/models/creditProductResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Manyak Server API
 * Manyak backend API documentation
 * OpenAPI spec version: v1
 */

export interface CreditProductResponse {
  productId?: string;
  baseCredits?: number;
  bonusCredits?: number;
  totalCredits?: number;
  webPriceKrw?: number;
  appPriceKrw?: number;
}