// src/api/generated/models/trialsResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Manyak Server API
 * Manyak backend API documentation
 * OpenAPI spec version: v1
 */
import type { TrialUsage } from './trialUsage';

export interface TrialsResponse {
  chatTurn?: TrialUsage;
  chatImage?: TrialUsage;
  storyCreation?: TrialUsage;
  storylineGeneration?: TrialUsage;
}